created plus button/social & comment modal
